requirements.txt
imutils==0.5.4
numpy==1.19.3
matplotlib==3.3.4
imageio==2.9.0
opencv_contrib_python==4.5.3.56
Flask==1.1.2
moviepy==1.0.3
absl_py==0.13.0
mediapipe==0.8.3
scipy==1.5.4
Pillow
PySimpleGUI==4.47.0